+++ OnlineHostelMgmt/src/main/java/com/cg/dao/IWardenDao.java
@@ -10,6 +10,6 @@ import com.cg.entities.Warden;
 @Repository
 public interface IWardenDao extends JpaRepository<Warden, Integer> {
 
-	public List<Warden> findByHostelId(Long hostelId);
+	public List<Warden> findByHostelId(Integer hostelId);
 
 }
